chore: refactor non-API views into separate modules/files
