+++ bin/upgrade_containers.sh
@@ -56,7 +56,7 @@ fi
 
 sudo -E docker compose \
     -f /home/${USER}/screenly/docker-compose.yml \
-    pull
+    build
 
 if [ -f /var/run/reboot-required ]; then
     exit 0
